Converted NodeEditor
