+++ package.json
@@ -45,6 +45,7 @@
     "@testing-library/dom": "^7.20.0",
     "@testing-library/react": "^10.4.3",
     "@types/d3-shape": "^3.0.2",
+    "@types/lodash": "^4.14.182",
     "@types/react": "^18.0.5",
     "@types/react-portal": "^4.0.4",
     "@typescript-eslint/eslint-plugin": "2.x",
